requirements.txt
Django==5.1
django-cors-headers==4.6.0
django-rest-framework==0.1.0
drf-yasg==1.21.7
pytest==8.3.4
pytest-django==4.9.0
psycopg2==2.9.9